+++ src/main/java/io/github/tufkan1/projectex/content/ProjectEXMenus.java
@@ -2,6 +2,7 @@ package io.github.tufkan1.projectex.content;
 
 import io.github.tufkan1.projectex.ProjectEX;
 import io.github.tufkan1.projectex.menu.TransmutationMenu;
+import io.github.tufkan1.projectex.menu.EmcMachineMenu;
 import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.world.flag.FeatureFlags;
@@ -13,6 +14,11 @@ public final class ProjectEXMenus {
         BuiltInRegistries.MENU,
         ProjectEX.id("transmutation"),
         new MenuType<>(TransmutationMenu::new, FeatureFlags.VANILLA_SET)
+    );
+    public static final MenuType<EmcMachineMenu> EMC_MACHINE = Registry.register(
+        BuiltInRegistries.MENU,
+        ProjectEX.id("emc_machine"),
+        new MenuType<>(EmcMachineMenu::new, FeatureFlags.VANILLA_SET)
     );
 
     private ProjectEXMenus() {